Splitting out the contents
The monolithic file was proving difficult to navigate and so break this out to improve matters.
